+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
+## v0.1.0 (2024-10-11)
 
+Initial Commit for resource-catalogue-fastapi user service
+- Add FastApi application to allow users to add items to catalogues through pulsar messages
+- Add middleware to integrate with OPA authentication to restrict user access